+++ crates/punktf-lib/src/profile/mod.rs
@@ -1,24 +1,72 @@
 //! Defines profiles and ways to layer multiple of them.
 
+pub mod dotfile;
+pub mod hook;
+pub mod link;
+pub mod source;
+pub mod transform;
+pub mod variables;
+
 use std::collections::{HashMap, HashSet};
-use std::ffi::OsString;
 use std::fs::File;
 use std::ops::Deref;
 use std::path::{Path, PathBuf};
 
-use color_eyre::eyre::{eyre, Context};
+use color_eyre::eyre::{bail, eyre, Context};
 use color_eyre::Result;
 use serde::{Deserialize, Serialize};
 
-use crate::hook::Hook;
-use crate::transform::ContentTransformer;
-use crate::variables::{Variables, Vars};
-use crate::{Dotfile, PunktfSource};
+use crate::profile::hook::Hook;
+use crate::profile::link::Symlink;
+use crate::profile::transform::ContentTransformer;
+use crate::profile::variables::{Variables, Vars};
+use crate::profile::{dotfile::Dotfile, source::PunktfSource};
+
+/// This enum represents all available merge modes `punktf` supports. The merge
+/// mode is important when a file already exists at the target location of a
+/// [`Dotfile`](`crate::profile::dotfile::Dotfile`).
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub enum MergeMode {
+	/// Overwrites the existing file.
+	Overwrite,
+
+	/// Keeps the existing file.
+	Keep,
+
+	/// Asks the user for input to decide what to do.
+	Ask,
+}
+
+impl Default for MergeMode {
+	fn default() -> Self {
+		Self::Overwrite
+	}
+}
+
+/// This struct represents the priority a
+/// [`Dotfile`](`crate::profile::dotfile::Dotfile`)
+/// can have. A bigger value means a higher priority. Dotfiles with lower priority
+/// won't be able to overwrite already deployed dotfiles with a higher one.
+#[derive(
+	Default, Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize,
+)]
+pub struct Priority(pub u32);
+
+impl Priority {
+	/// Creates a new instance with the given `priority`.
+	pub const fn new(priority: u32) -> Self {
+		Self(priority)
+	}
+}
 
 /// A profile is a collection of dotfiles and variables, options and hooks.
 #[derive(Default, Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
 pub struct Profile {
+	/// Aliases for this profile which can be used instead of the file name.
+	#[serde(skip_serializing_if = "Vec::is_empty", default)]
+	pub aliases: Vec<String>,
+
 	/// Defines the base profile. All settings from the base are merged with the
 	/// current profile. The settings from the current profile take precendence.
 	/// Dotfiles are merged on the dotfile level (not specific dotfile settings level).
@@ -34,7 +82,8 @@ pub struct Profile {
 	pub transformers: Vec<ContentTransformer>,
 
 	/// Target root path of the deployment. Will be used as file stem for the dotfiles
-	/// when not overwritten by [`Dotfile::overwrite_target`].
+	/// when not overwritten by
+	/// [`Dotfile::overwrite_target`](`crate::profile::dotfile::Dotfile::overwrite_target`).
 	#[serde(skip_serializing_if = "Option::is_none", default)]
 	pub target: Option<PathBuf>,
 
@@ -50,6 +99,10 @@ pub struct Profile {
 	/// Dotfiles which will be deployed.
 	#[serde(skip_serializing_if = "Vec::is_empty", default)]
 	pub dotfiles: Vec<Dotfile>,
+
+	/// Symlinks which will be deployed.
+	#[serde(rename = "links", skip_serializing_if = "Vec::is_empty", default)]
+	pub symlinks: Vec<Symlink>,
 }
 
 impl Profile {
@@ -127,8 +180,8 @@ impl Profile {
 #[derive(Default, Debug, Clone, PartialEq, Eq)]
 pub struct LayeredVariables {
 	/// Stores the variables together with the index, which indexed
-	/// [`LayeredProfile::profile_names`] to retrieve the name of the profile,
-	/// the variable came from.
+	/// [`LayeredProfile::profile_names`](`crate::profile::LayeredProfile::profile_names`)
+	/// to retrieve the name of the profile, the variable came from.
 	pub inner: HashMap<String, (usize, String)>,
 }
 
@@ -145,31 +198,39 @@ impl Vars for LayeredVariables {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct LayeredProfile {
 	/// All names of the profile which where collected from the extend chain.
-	profile_names: Vec<String>,
+	pub profile_names: Vec<String>,
 
 	/// The target of the deployment.
 	///
 	/// This is the first value found by traversing the extend chain from the
 	/// top.
-	target: Option<(usize, PathBuf)>,
+	pub target: Option<(usize, PathBuf)>,
 
 	/// The variables collected from all profiles of the extend chain.
-	variables: LayeredVariables,
+	pub variables: LayeredVariables,
 
 	/// The content transformer collected from all profiles of the extend chain.
-	transformers: Vec<(usize, ContentTransformer)>,
+	pub transformers: Vec<(usize, ContentTransformer)>,
 
 	/// The pre-hooks collected from all profiles of the extend chain.
-	pre_hooks: Vec<(usize, Hook)>,
+	pub pre_hooks: Vec<(usize, Hook)>,
 
 	/// The post-hooks collected from all profiles of the extend chain.
-	post_hooks: Vec<(usize, Hook)>,
+	pub post_hooks: Vec<(usize, Hook)>,
 
 	/// The dotfiles collected from all profiles of the extend chain.
 	///
-	/// The index indexes into [`LayeredProfile::profile_names`] to retrieve
-	/// the name of the profile from which the dotfile came from.
-	dotfiles: Vec<(usize, Dotfile)>,
+	/// The index indexes into
+	/// [`LayeredProfile::profile_names`](`crate::profile::LayeredProfile::profile_names`)
+	/// to retrieve the name of the profile from which the dotfile came from.
+	pub dotfiles: Vec<(usize, Dotfile)>,
+
+	/// The symlinks collected from all profiles of the extend chain.
+	///
+	/// The index indexes into
+	/// [`LayeredProfile::profile_names`](`crate::profile::LayeredProfile::profile_names`)
+	/// to retrieve the name of the profile from which the link came from.
+	pub symlinks: Vec<(usize, Symlink)>,
 }
 
 impl LayeredProfile {
@@ -179,7 +240,7 @@ impl LayeredProfile {
 	}
 
 	/// Returns the target path for the profile together with the index into
-	/// [`LayeredProfile::profile_names`].
+	/// [`LayeredProfile::profile_names`](`crate::profile::LayeredProfile::profile_names`).
 	pub fn target(&self) -> Option<(&str, &Path)> {
 		self.target
 			.as_ref()
@@ -219,6 +280,11 @@ impl LayeredProfile {
 	/// Returns all collected dotfiles for the profile.
 	pub fn dotfiles(&self) -> impl Iterator<Item = &Dotfile> {
 		self.dotfiles.iter().map(|(_, dotfile)| dotfile)
+	}
+
+	/// Returns all collected symlinks for the profile.
+	pub fn symlinks(&self) -> impl Iterator<Item = &Symlink> {
+		self.symlinks.iter().map(|(_, symlink)| symlink)
 	}
 }
 
@@ -283,31 +349,31 @@ impl LayeredProfileBuilder {
 			}
 		}
 
-		let mut pre_hooks = Vec::new();
-
-		for (idx, hooks) in self
+		let pre_hooks = self
 			.profiles
 			.iter()
 			.enumerate()
-			.map(|(idx, profile)| (idx, &profile.pre_hooks))
-		{
-			for hook in hooks.iter().cloned() {
-				pre_hooks.push((idx, hook));
-			}
-		}
+			.flat_map(|(idx, profile)| {
+				profile
+					.pre_hooks
+					.iter()
+					.cloned()
+					.map(move |hook| (idx, hook))
+			})
+			.collect();
 
-		let mut post_hooks = Vec::new();
-
-		for (idx, hooks) in self
+		let post_hooks = self
 			.profiles
 			.iter()
 			.enumerate()
-			.map(|(idx, profile)| (idx, &profile.post_hooks))
-		{
-			for hook in hooks.iter().cloned() {
-				post_hooks.push((idx, hook));
-			}
-		}
+			.flat_map(|(idx, profile)| {
+				profile
+					.post_hooks
+					.iter()
+					.cloned()
+					.map(move |hook| (idx, hook))
+			})
+			.collect();
 
 		let mut added_dotfile_paths = HashSet::new();
 		let mut dotfiles = Vec::new();
@@ -326,6 +392,19 @@ impl LayeredProfileBuilder {
 			}
 		}
 
+		let symlinks = self
+			.profiles
+			.iter()
+			.enumerate()
+			.flat_map(|(idx, profile)| {
+				profile
+					.symlinks
+					.iter()
+					.cloned()
+					.map(move |link| (idx, link))
+			})
+			.collect();
+
 		LayeredProfile {
 			profile_names: self.profile_names,
 			target,
@@ -334,8 +413,119 @@ impl LayeredProfileBuilder {
 			pre_hooks,
 			post_hooks,
 			dotfiles,
+			symlinks,
 		}
 	}
+}
+
+/// A minimal struct to read the `aliases` from a profile file.
+///
+/// This is used for profile name resolution.
+#[derive(Default, Debug, Serialize, Deserialize)]
+#[serde(default)]
+struct Aliases {
+	/// Aliases of a profile.
+	///
+	/// These can be used in place of the profile name for cli and extend resolution.
+	aliases: Vec<String>,
+}
+
+/// Collects all profile names and aliases from the `profiles` directory.
+pub fn collect_profile_names(source: &PunktfSource) -> Result<HashMap<String, PathBuf>> {
+	log::info!("Collecting profile names and aliases");
+
+	/// Tries to read all alias from a given file.
+	fn get_aliases(path: &Path, extension: &str) -> Option<Aliases> {
+		let Ok(file) = File::open(path) else {
+			log::debug!("[{}] Failed to read content", path.display());
+				return None;
+			};
+
+		#[cfg(feature = "profile-json")]
+		{
+			if extension.eq_ignore_ascii_case("json") {
+				let Ok(aliases) = serde_json::from_reader(file) else {
+						log::debug!("[{}] Failed to read aliases", path.display());
+						return None;
+					};
+
+				return Some(aliases);
+			}
+		}
+
+		#[cfg(feature = "profile-yaml")]
+		{
+			if extension.eq_ignore_ascii_case("yaml") || extension.eq_ignore_ascii_case("yml") {
+				let Ok(aliases) = serde_yaml::from_reader(file) else {
+						log::debug!("[{}] Failed to read aliases", path.display());
+						return None;
+					};
+
+				return Some(aliases);
+			}
+		}
+
+		None
+	}
+
+	let mut names = HashMap::new();
+
+	let dents = source.profiles().read_dir()?;
+	for dent in dents {
+		let dent = dent?;
+		let path = dent.path();
+
+		let Ok(ft) = dent.file_type() else {
+			log::debug!("[{}] Failed to get file type", path.display());
+			continue;
+		};
+
+		if !ft.is_file() {
+			log::debug!("[{}] Not a file", path.display());
+			continue;
+		}
+
+		let Some(extension) = path.extension().and_then(|e| e.to_str()) else {
+			log::debug!("[{}] Failed to get file extension", path.display());
+			continue;
+		};
+
+		let Some(name) = path.file_name().and_then(|n| n.to_str()) else {
+			log::debug!("[{}] Failed to get file name", path.display());
+			continue;
+		};
+		// Remove extension and `.`
+		let name = &name[..(name.len() - (extension.len() + 1))];
+
+		for alias in get_aliases(&path, extension)
+			.into_iter()
+			.flat_map(|a| a.aliases.into_iter())
+		{
+			log::debug!("[{}] Adding alias {}", path.display(), alias);
+
+			if let Some(evicted) = names.insert(alias.clone(), path.clone()) {
+				bail!(
+					"[{}] The profile alias {} is already taken by {}",
+					path.display(),
+					alias,
+					evicted.display()
+				);
+			}
+		}
+
+		if let Some(evicted) = names.insert(name.to_string(), path.clone()) {
+			bail!(
+				"[{}] The profile name {} is already taken by {}",
+				path.display(),
+				name,
+				evicted.display()
+			);
+		}
+	}
+
+	log::info!("Found {} profile names and aliases", names.len());
+
+	Ok(names)
 }
 
 /// Recursively resolves a profile and it's [extend
@@ -345,45 +535,58 @@ pub fn resolve_profile(
 	builder: &mut LayeredProfileBuilder,
 	source: &PunktfSource,
 	name: &str,
-	resolved_profiles: &mut Vec<OsString>,
 ) -> Result<()> {
-	log::trace!("Resolving profile `{}`", name);
+	/// Recursive resolution of all profiles needed.
+	///
+	/// Checks for cycles while resolving.
+	fn _resolve_profile_inner(
+		profiles: &HashMap<String, PathBuf>,
+		builder: &mut LayeredProfileBuilder,
+		name: &str,
+		resolved_profiles: &mut Vec<String>,
+	) -> Result<()> {
+		log::trace!("Resolving profile `{}`", name);
 
-	let path = source.find_profile_path(name)?;
-	let file_name = path
-		.file_name()
-		.unwrap_or_else(|| panic!("Profile path has no file name ({:?})", path))
-		.to_os_string();
+		let path = profiles
+			.get(name)
+			.ok_or_else(|| eyre!("No profile found for name {}", name))?;
 
-	let mut profile = Profile::from_file(&path)?;
+		let mut profile = Profile::from_file(path)?;
+		let name = name.to_string();
 
-	if !profile.extends.is_empty() && resolved_profiles.contains(&file_name) {
-		// profile was already resolve and has "children" which will lead to
-		// a loop while resolving
-		return Err(eyre!(
+		if !profile.extends.is_empty() && resolved_profiles.contains(&name) {
+			// profile was already resolve and has "children" which will lead to
+			// a loop while resolving
+			return Err(eyre!(
 			"Circular dependency detected while parsing `{}` (required by: `{:?}`) (Stack: {:#?})",
 			name,
 			resolved_profiles.last(),
 			resolved_profiles
 		));
+		}
+
+		let mut extends = Vec::new();
+		std::mem::swap(&mut extends, &mut profile.extends);
+
+		builder.add(name.clone(), profile);
+
+		resolved_profiles.push(name);
+
+		for child in extends {
+			_resolve_profile_inner(profiles, builder, &child, resolved_profiles)?;
+		}
+
+		let _ = resolved_profiles
+			.pop()
+			.expect("Misaligned push/pop operation");
+
+		Ok(())
 	}
 
-	let mut extends = Vec::new();
-	std::mem::swap(&mut extends, &mut profile.extends);
+	let available_profiles = collect_profile_names(source)?;
+	let mut resolved_profiles = Vec::new();
 
-	builder.add(name.to_string(), profile);
-
-	resolved_profiles.push(file_name);
-
-	for child in extends {
-		resolve_profile(builder, source, &child, resolved_profiles)?;
-	}
-
-	let _ = resolved_profiles
-		.pop()
-		.expect("Misaligned push/pop operation");
-
-	Ok(())
+	_resolve_profile_inner(&available_profiles, builder, name, &mut resolved_profiles)
 }
 
 #[cfg(test)]
@@ -391,10 +594,19 @@ mod tests {
 	use std::collections::HashMap;
 
 	use super::*;
-	use crate::hook::Hook;
+	use crate::profile::hook::Hook;
+	use crate::profile::variables::Variables;
 	use crate::profile::Profile;
-	use crate::variables::Variables;
-	use crate::{MergeMode, Priority};
+	use crate::profile::{MergeMode, Priority};
+
+	#[test]
+	fn priority_order() {
+		crate::tests::setup_test_env();
+
+		assert!(Priority::default() == Priority::new(0));
+		assert!(Priority::new(0) == Priority::new(0));
+		assert!(Priority::new(2) > Priority::new(1));
+	}
 
 	#[test]
 	#[cfg(feature = "profile-json")]
@@ -411,6 +623,7 @@ mod tests {
 
 		let profile = Profile {
 			extends: Vec::new(),
+			aliases: vec![],
 			variables: Some(Variables {
 				inner: profile_vars,
 			}),
@@ -442,6 +655,7 @@ mod tests {
 					template: Some(false),
 				},
 			],
+			symlinks: vec![],
 		};
 
 		let json = serde_json::to_string(&profile).expect("Profile to be serializeable");